implement review comments (file renames)
