+++ src/vs/workbench/contrib/chat/browser/actions/reusablePromptActions/dialogs/askToSelectPrompt/askToSelectPrompt.ts
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { IChatWidget } from '../../../../chat.js';
-import { attachInstructionsFiles } from './utils/attachPrompt.js';
+import { attachInstructionsFiles } from './utils/attachInstructions.js';
 import { handleButtonClick } from './utils/handleButtonClick.js';
 import { URI } from '../../../../../../../../base/common/uri.js';
 import { createPromptPickItem } from './utils/createPromptPickItem.js';
