+++ src/vs/workbench/contrib/chat/browser/actions/reusablePromptActions/dialogs/askToSelectPrompt/utils/attachInstructions.ts
@@ -6,12 +6,11 @@
 import { IChatWidget, showChatView } from '../../../../../chat.js';
 import { URI } from '../../../../../../../../../base/common/uri.js';
 import { ACTION_ID_NEW_CHAT } from '../../../../chatClearActions.js';
-import { extUri } from '../../../../../../../../../base/common/resources.js';
 import { assertDefined } from '../../../../../../../../../base/common/types.js';
-import { IChatAttachInstructionsActionOptions } from '../../../chatAttachPromptAction.js';
+import { IChatAttachInstructionsActionOptions } from '../../../chatAttachInstructionsAction.js';
 import { IViewsService } from '../../../../../../../../services/views/common/viewsService.js';
 import { ICommandService } from '../../../../../../../../../platform/commands/common/commands.js';
-import { detachPrompt } from './detachPrompt.js';
+import { IRunPromptOptions } from './runPrompt.js';
 
 /**
  * Options for the {@link attachInstructionsFiles} function.
@@ -39,64 +38,7 @@ interface IAttachResult {
 	readonly wereAlreadyAttached: readonly URI[];
 }
 
-/**
- * Options for the {@link runPromptFile} function.
- */
-export interface IRunPromptOptions {
-	/**
-	 * Chat widget instance to attach the prompt to.
-	 */
-	readonly widget?: IChatWidget;
-	/**
-	 * Whether to create a new chat session and
-	 * attach the instructions file to it.
-	 */
-	readonly inNewChat?: boolean;
 
-	readonly viewsService: IViewsService;
-	readonly commandService: ICommandService;
-}
-
-
-/**
- * Return value of the {@link runPromptFile} function.
- */
-interface IRunPromptResult {
-	readonly widget: IChatWidget;
-}
-
-/**
- * Check if provided uri is already attached to chat
- * input as an implicit  "current file" context.
- */
-const isAttachedAsCurrentPrompt = (
-	promptUri: URI,
-	widget: IChatWidget,
-): boolean => {
-	const { implicitContext } = widget.input;
-	if (implicitContext === undefined) {
-		return false;
-	}
-
-	if (implicitContext.isInstructions === false) {
-		return false;
-	}
-
-	if (implicitContext.enabled === false) {
-		return false;
-	}
-
-	assertDefined(
-		implicitContext.value,
-		'Prompt value must always be defined.',
-	);
-
-	const uri = URI.isUri(implicitContext.value)
-		? implicitContext.value
-		: implicitContext.value.uri;
-
-	return extUri.isEqual(promptUri, uri);
-};
 
 /**
  * Attaches provided instructions to a chat input.
@@ -121,42 +63,12 @@ export const attachInstructionsFiles = async (
 };
 
 /**
- * Runs the prompt file.
- */
-export const runPromptFile = async (
-	file: URI,
-	options: IRunPromptOptions,
-): Promise<IRunPromptResult> => {
-
-	const widget = await getChatWidgetObject(options);
-
-	let wasAlreadyAttached = true;
-	if (isAttachedAsCurrentPrompt(file, widget) === false) {
-		wasAlreadyAttached = widget
-			.attachmentModel
-			.promptInstructions
-			.add(file);
-	}
-
-	// submit the prompt immediately
-	await widget.acceptInput();
-
-	// detach the prompt immediately, unless was attached
-	// before the action was executed
-	if (wasAlreadyAttached === false) {
-		await detachPrompt(file, { widget });
-	}
-
-	return { widget };
-};
-
-/**
  * Gets a chat widget based on the provided {@link IChatAttachInstructionsActionOptions.widget widget}
  * reference and the `inNewChat` flag.
  *
  * @throws if failed to reveal a chat widget.
  */
-const getChatWidgetObject = async (
+export const getChatWidgetObject = async (
 	options: IAttachOptions | IRunPromptOptions,
 ): Promise<IChatWidget> => {
 	const { widget, inNewChat } = options;
