+++ src/vs/workbench/contrib/chat/browser/actions/reusablePromptActions/index.ts
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { registerRunPromptActions } from './chatRunPromptAction.js';
-import { registerAttachPromptActions } from './chatAttachPromptAction.js';
+import { registerAttachPromptActions } from './chatAttachInstructionsAction.js';
 
 /**
  * Helper to register all actions related to reusable prompt files.
@@ -14,4 +14,4 @@ export const registerReusablePromptActions = () => {
 	registerAttachPromptActions();
 };
 
-export { runAttachInstructionsAction } from './chatAttachPromptAction.js';
+export { runAttachInstructionsAction } from './chatAttachInstructionsAction.js';
