+++ src/vs/workbench/contrib/chat/browser/promptSyntax/contributions/ attachInstructionsCommand.ts
